+++ src/config.js
@@ -1,31 +1,14 @@
+const { Regex } = require("@companion-module/base");
+const utils = require("./utils");
+const actions = require("./actions");
+
 module.exports = {
-	/**
-	 * The user updated the config.
-	 *
-	 * @param config         The new config object
-	 */
-	updateConfig: function(config) {
-		let self = this;
-
-		// Reconnect to Pandoras Box if the IP changed
-		if (self.config.host !== config.host || self.isConnected() === false) {
-			self.config.host = config.host;
-			self.init_tcp();
-		}
-
-		// Keep config around
-		self.config = config;
-
-		// Build actions
-		self.actions();
-	},
-
  /**
  * Return config fields for web config.
  *
  * @returns      The config fields for the module {Object}
  */
-	getConfigFields: function() {
+	getConfigFields() {
 		let self = this;
 		return [
 			{
@@ -40,15 +23,21 @@ module.exports = {
 				id: 'host',
 				label: 'Target IP',
 				width: 6,
-				regex: self.REGEX_IP
+				regex: Regex.IP
+			},
+			{
+				type: 'static-text',
+				id: 'port',
+				label: 'Port (static): ',
+				value: '6211'
 			},
 			{
 				type: 'textinput',
 				id: 'domain',
 				width: 5,
 				label: 'Domain',
-				value: "0",
-				regex: self.REGEX_NUMBER
+				default: "0",
+				regex: Regex.NUMBER
 			}
 		];
 	}
